Shorthand lib format; imported manifest flag
For repetitive independent project imports, you can use the new "concat"
shorthand representation, which is an array. This allows you to replace:
	{reqs: ["jquery.loc", "jquery.api", "jquery.foo", "jquery.bar"]}
with:
	{reqs: ["jquery.", ["api", "foo", "bar"]}

This more compact format simply appends the first element of the array
to each element in the second array.

Also, manifests imported from the CDN will now get an "autogenerated"
flag to identify them as "can be replaced by an import".

This should prevent a user having a custom manifest smashed by pulling
down a cdnjs resource of the same name.

Change-Id: Ib7a7cc1a5796fd06275a5f2a6590a875e2f99976